refactor(aionrs): split windows job helper
Move Windows Job Object FFI out of the command runner module so runner lifecycle code stays focused on command execution.
