+++ crates/aion-process/src/lib.rs
@@ -1,4 +1,6 @@
 mod command_runner;
+#[cfg(windows)]
+mod windows_job;
 
 pub use command_runner::{
     CommandRunResult, CommandRunner, DEFAULT_POST_PROCESS_DRAIN, DEFAULT_TIMEOUT,
